Create new layer store
