+++ web/src/components/sidebars/DatasetsPanel.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import DatasetList from '@/components/DatasetList.vue'
 import DetailView from '@/components/DetailView.vue'
-import { useMapStore } from '@/store/map';
+import { useLayerStore } from '@/store/layer';
 import { Dataset, Layer } from '@/types';
 
 
@@ -12,7 +12,7 @@ const props = defineProps<{
 function toggleSelected(items: Layer[]) {
   items.forEach((item) => {
     const layer = item as Layer;
-    useMapStore().addLayer(layer);
+    useLayerStore().addLayer(layer);
   })
 }
 </script>
